build universal static libraries, add them to bin directory

+++ samples/TTCatalog/Classes/ScrollViewTestController.m
@@ -109,7 +109,7 @@
 #pragma mark TTPageControl
 
 - (IBAction)changePage:(id)sender {
-  int page = _pageControl.currentPage;
+  NSInteger page = _pageControl.currentPage;
   [_scrollView setCenterPageIndex:page];
 }
